+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Versioning practices: [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
  - Fix 'wult/ndl' not generating statistics tabs in HTML reports for some
    results with format version '1.2'.
 ### Added
+ - ndl start: support 'local' and 'remote' values for the '--cpunum' option.
 ### Removed
 ### Changed